[hackathon] improve OS support (#344)
* [embed] making JMX and APM modular.

* [multiplatform] multiple file refactor to enable freebsd, solaris, netbsd builds.

* [host] put linux in nix bundle for OS version.

* [jmx] selectively run tests according to build.

* [metadata][v5] lets not neglect darwin here :)

* [rakefile] small refactor to avoid duplication.

* [rakefile] document puppy option in task descs.
