+++ pkg/collector/corechecks/embed/jmx_test.go
@@ -1,3 +1,5 @@
+// +build jmx
+
 package embed
 
 import (
