+++ pkg/collector/corechecks/embed/jmxloader.go
@@ -1,3 +1,5 @@
+// +build jmx
+
 package embed
 
 import (
