+++ pkg/config/config_darwin.go
@@ -4,4 +4,5 @@ const (
 	defaultConfdPath            = "/opt/datadog-agent/etc/conf.d"
 	defaultAdditionalChecksPath = "/opt/datadog-agent/etc/checks.d"
 	defaultLogPath              = "/var/log/datadog/agent.log"
+	defaultJMXPipePath          = "/opt/datadog-agent/run"
 )
