--- pkg/config/config_linux.go
@@ -1,7 +1,0 @@
-package config
-
-const (
-	defaultConfdPath            = "/etc/dd-agent/conf.d"
-	defaultAdditionalChecksPath = "/etc/dd-agent/checks.d"
-	defaultLogPath              = "/var/log/datadog/agent.log"
-)
