+++ pkg/metadata/resources/resources_other.go
@@ -1,0 +1,10 @@
+// +build freebsd netbsd openbsd solaris dragonfly
+
+package resources
+
+// GetPayload currently just a stub.
+func GetPayload(hostname string) *Payload {
+
+	//unimplemented for misc platforms.
+	return &Payload{}
+}
