+++ pkg/metadata/v5/payload.go
@@ -2,7 +2,6 @@ package v5
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/metadata/common"
-	"github.com/DataDog/datadog-agent/pkg/metadata/gohai"
 	"github.com/DataDog/datadog-agent/pkg/metadata/host"
 	"github.com/DataDog/datadog-agent/pkg/metadata/resources"
 )
@@ -20,20 +19,4 @@ type HostPayload struct {
 // ResourcesPayload wraps Payload from the resources package
 type ResourcesPayload struct {
 	resources.Payload
-}
-
-// GohaiPayload wraps Payload from the gohai package
-type GohaiPayload struct {
-	gohai.Payload
-}
-
-// Payload handles the JSON unmarshalling of the metadata payload
-type Payload struct {
-	CommonPayload
-	HostPayload
-	ResourcesPayload
-	// TODO: host-tags
-	// TODO: external_host_tags
-	GohaiPayload
-	// TODO: agent_checks
 }
